# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Personal Finance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: getTransactions
      tags: [transactions]
      summary: List transactions with optional filters
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
            description: "Filter by month YYYY-MM"
        - name: category_id
          in: query
          required: false
          schema:
            type: integer
        - name: account_id
          in: query
          required: false
          schema:
            type: integer
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [income, expense, transfer]
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/import:
    post:
      operationId: importTransactions
      tags: [transactions]
      summary: Import transactions from CSV
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportTransactionsRequest"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        "200":
          description: Transaction updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date
        description:
          type: string
        account_id:
          type: integer
        category_id:
          type: integer
          nullable: true
        subcategory_id:
          type: integer
          nullable: true
        amount:
          type: number
          description: Negative = expense, positive = income
        person:
          type: string
        type:
          type: string
          enum: [income, expense, transfer]
        account_name:
          type: string
          nullable: true
        category_name:
          type: string
          nullable: true
        subcategory_name:
          type: string
          nullable: true
      required:
        - id
        - date
        - description
        - account_id
        - amount
        - person
        - type
    CreateTransactionRequest:
      type: object
      properties:
        date:
          type: string
          format: date
        description:
          type: string
        account_id:
          type: integer
        category_id:
          type: integer
          nullable: true
        subcategory_id:
          type: integer
          nullable: true
        amount:
          type: number
        person:
          type: string
        type:
          type: string
          enum: [income, expense, transfer]
      required:
        - date
        - description
        - account_id
        - amount
        - person
        - type
    ImportTransactionsRequest:
      type: object
      properties:
        csv_data:
          type: string
          description: Raw CSV string
      required:
        - csv_data
    ImportResult:
      type: object
      properties:
        imported:
          type: integer
        errors:
          type: array
          items:
            type: string
      required:
        - imported
        - errors
...